+++ source/mysql/result.d
@@ -247,15 +247,6 @@ public:
 	}
 }
 
-/// Deprecated. Replaced by
-/// $(LINK2 https://dlang.org/phobos/std_variant.html, std.variant.Variant).
-deprecated("Use std.variant.Variant instead.")
-struct DBValue
-{
-	Variant value;
-	bool isNull;
-}
-
 /++
 A $(LINK2 http://dlang.org/phobos/std_range_primitives.html#isRandomAccessRange, random access range)
 of Row.